+++ lib/src/account/mod.rs
@@ -1,6 +1,11 @@
+mod close;
 mod cpi;
-mod traits;
+mod deserialize;
+mod lamports;
 mod validation;
 
+pub use close::*;
 pub use cpi::*;
-pub use traits::*;
+pub use deserialize::*;
+pub use lamports::*;
+pub use validation::*;
