+++ lib/src/account/validation.rs
@@ -1,10 +1,22 @@
 use bytemuck::Pod;
 use solana_program::{account_info::AccountInfo, program_error::ProgramError, pubkey::Pubkey};
 
-use crate::{
-    trace, AccountDeserialize, AccountInfoValidation, AsAccount, CloseAccount, Discriminator,
-    LamportTransfer,
-};
+use crate::trace;
+
+use super::{AccountDeserialize, Discriminator};
+
+pub trait AccountInfoValidation {
+    fn is_signer(&self) -> Result<&Self, ProgramError>;
+    fn is_writable(&self) -> Result<&Self, ProgramError>;
+    fn is_executable(&self) -> Result<&Self, ProgramError>;
+    fn is_empty(&self) -> Result<&Self, ProgramError>;
+    fn is_type<T: Discriminator>(&self, program_id: &Pubkey) -> Result<&Self, ProgramError>;
+    fn is_program(&self, program_id: &Pubkey) -> Result<&Self, ProgramError>;
+    fn is_sysvar(&self, sysvar_id: &Pubkey) -> Result<&Self, ProgramError>;
+    fn has_address(&self, address: &Pubkey) -> Result<&Self, ProgramError>;
+    fn has_owner(&self, program_id: &Pubkey) -> Result<&Self, ProgramError>;
+    fn has_seeds(&self, seeds: &[&[u8]], program_id: &Pubkey) -> Result<&Self, ProgramError>;
+}
 
 impl AccountInfoValidation for AccountInfo<'_> {
     #[track_caller]
@@ -109,6 +121,20 @@ impl AccountInfoValidation for AccountInfo<'_> {
     }
 }
 
+/// Performs:
+/// 1. Program owner check
+/// 2. Discriminator byte check
+/// 3. Checked bytemuck conversion of account data to &T or &mut T.
+pub trait AsAccount {
+    fn as_account<T>(&self, program_id: &Pubkey) -> Result<&T, ProgramError>
+    where
+        T: AccountDeserialize + Discriminator + Pod;
+
+    fn as_account_mut<T>(&self, program_id: &Pubkey) -> Result<&mut T, ProgramError>
+    where
+        T: AccountDeserialize + Discriminator + Pod;
+}
+
 impl AsAccount for AccountInfo<'_> {
     #[track_caller]
     fn as_account<T>(&self, program_id: &Pubkey) -> Result<&T, ProgramError>
@@ -156,30 +182,32 @@ impl AsAccount for AccountInfo<'_> {
     }
 }
 
-impl<'a, 'info> LamportTransfer<'a, 'info> for AccountInfo<'info> {
-    #[inline(always)]
-    fn send(&'a self, lamports: u64, to: &'a AccountInfo<'info>) {
-        **self.lamports.borrow_mut() -= lamports;
-        **to.lamports.borrow_mut() += lamports;
-    }
+pub trait AccountValidation {
+    fn assert<F>(&self, condition: F) -> Result<&Self, ProgramError>
+    where
+        F: Fn(&Self) -> bool;
 
-    #[inline(always)]
-    fn collect(&'a self, lamports: u64, from: &'a AccountInfo<'info>) -> Result<(), ProgramError> {
-        solana_program::program::invoke(
-            &solana_program::system_instruction::transfer(from.key, self.key, lamports),
-            &[from.clone(), self.clone()],
-        )
-    }
-}
+    fn assert_err<F>(&self, condition: F, err: ProgramError) -> Result<&Self, ProgramError>
+    where
+        F: Fn(&Self) -> bool;
 
-impl<'a, 'info> CloseAccount<'a, 'info> for AccountInfo<'info> {
-    fn close(&'a self, to: &'a AccountInfo<'info>) -> Result<(), ProgramError> {
-        // Realloc data to zero.
-        self.realloc(0, true)?;
+    fn assert_msg<F>(&self, condition: F, msg: &str) -> Result<&Self, ProgramError>
+    where
+        F: Fn(&Self) -> bool;
 
-        // Return rent lamports.
-        self.send(self.lamports(), to);
+    fn assert_mut<F>(&mut self, condition: F) -> Result<&mut Self, ProgramError>
+    where
+        F: Fn(&Self) -> bool;
 
-        Ok(())
-    }
+    fn assert_mut_err<F>(
+        &mut self,
+        condition: F,
+        err: ProgramError,
+    ) -> Result<&mut Self, ProgramError>
+    where
+        F: Fn(&Self) -> bool;
+
+    fn assert_mut_msg<F>(&mut self, condition: F, msg: &str) -> Result<&mut Self, ProgramError>
+    where
+        F: Fn(&Self) -> bool;
 }
